Added an extension to convert a versionedreference to a non versioned reference Renamed ToReference to ToTypedReference for reference variants
