+++ src/MooVC.Architecture.Tests/Ddd/ReferenceExtensionsTests/WhenToTypedReferenceIsCalled.cs
@@ -3,13 +3,13 @@ namespace MooVC.Architecture.Ddd.ReferenceExtensionsTests
     using System;
     using Xunit;
 
-    public sealed class WhenToReferenceIsCalled
+    public sealed class WhenToTypedReferenceIsCalled
     {
         [Fact]
         public void GivenAMatchingReferenceThenNoExceptionIsThrown()
         {
             Reference generic = new Reference<AggregateRoot>(Guid.NewGuid());
-            var typed = generic.ToReference<AggregateRoot>();
+            Reference<AggregateRoot> typed = generic.ToTypedReference<AggregateRoot>();
 
             Assert.Same(generic, typed);
         }
@@ -18,7 +18,7 @@ namespace MooVC.Architecture.Ddd.ReferenceExtensionsTests
         public void GivenAMatchingEmptyReferenceThenNoExceptionIsThrown()
         {
             Reference generic = Reference<AggregateRoot>.Empty;
-            var typed = generic.ToReference<AggregateRoot>();
+            Reference<AggregateRoot> typed = generic.ToTypedReference<AggregateRoot>();
 
             Assert.Same(generic, typed);
         }
@@ -29,7 +29,7 @@ namespace MooVC.Architecture.Ddd.ReferenceExtensionsTests
             Reference reference = new Reference<EventCentricAggregateRoot>(Guid.NewGuid());
 
             ArgumentException exception = Assert.Throws<ArgumentException>(
-                () => reference.ToReference<AggregateRoot>());
+                () => reference.ToTypedReference<AggregateRoot>());
 
             Assert.Equal(nameof(reference), exception.ParamName);
         }
@@ -40,7 +40,7 @@ namespace MooVC.Architecture.Ddd.ReferenceExtensionsTests
             Reference reference = Reference<EventCentricAggregateRoot>.Empty;
 
             ArgumentException exception = Assert.Throws<ArgumentException>(
-                () => reference.ToReference<AggregateRoot>());
+                () => reference.ToTypedReference<AggregateRoot>());
 
             Assert.Equal(nameof(reference), exception.ParamName);
         }
