+++ src/MooVC.Architecture.Tests/Ddd/VersionedReferenceExtensionsTests/WhenToReferenceIsCalled.cs
@@ -6,43 +6,13 @@ namespace MooVC.Architecture.Ddd.VersionedReferenceExtensionsTests
     public sealed class WhenToReferenceIsCalled
     {
         [Fact]
-        public void GivenAMatchingVersionedReferenceThenNoExceptionIsThrown()
+        public void GivenAVersionedReferenceThenAReferenceIsReturned()
         {
-            VersionedReference generic = new VersionedReference<AggregateRoot>(Guid.NewGuid());
-            VersionedReference<AggregateRoot> typed = generic.ToReference<AggregateRoot>();
+            var versioned = new VersionedReference<AggregateRoot>(Guid.NewGuid());
+            var nonVersioned = versioned.ToReference();
 
-            Assert.Same(generic, typed);
-        }
-
-        [Fact]
-        public void GivenAMatchingEmptyVersionedReferenceThenNoExceptionIsThrown()
-        {
-            VersionedReference generic = VersionedReference<AggregateRoot>.Empty;
-            VersionedReference<AggregateRoot> typed = generic.ToReference<AggregateRoot>();
-
-            Assert.Same(generic, typed);
-        }
-
-        [Fact]
-        public void GivenAMismatchingVersionedReferenceThenAnArgumentExceptionIsThrown()
-        {
-            VersionedReference reference = new VersionedReference<EventCentricAggregateRoot>(Guid.NewGuid());
-
-            ArgumentException exception = Assert.Throws<ArgumentException>(
-                () => reference.ToReference<AggregateRoot>());
-
-            Assert.Equal(nameof(reference), exception.ParamName);
-        }
-
-        [Fact]
-        public void GivenAMismatchingEmptyVersionedReferenceThenAnArgumentExceptionIsThrown()
-        {
-            VersionedReference reference = VersionedReference<EventCentricAggregateRoot>.Empty;
-
-            ArgumentException exception = Assert.Throws<ArgumentException>(
-                () => reference.ToReference<AggregateRoot>());
-
-            Assert.Equal(nameof(reference), exception.ParamName);
+            Assert.True(versioned == nonVersioned);
+            Assert.NotSame(versioned, nonVersioned);
         }
     }
 }
